refactor(articles and news): articles and news preview generator
