+++ src/components/news-preview-small/index.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { useParams } from 'react-router-dom';
 import { TArticle } from 'utils/types/article';
 import CardArticlePreview from 'components/cards/article-preview/article-preview';
-import { newsExample } from './data/news';
+import { generateSmallPreview } from 'utils/previewSmall';
 
 import styles from './styles.module.scss';
 
@@ -14,19 +14,11 @@ interface INewsPreviewSmall {
 const maxNumCardsDesktop = 2;
 
 export const NewsPreviewSmall: FC<INewsPreviewSmall> = ({
-  data = newsExample,
+  data,
   route = '/',
 }) => {
-  const { id } = useParams();
-  const randomData: TArticle[] = [];
-
-  while (randomData.length < maxNumCardsDesktop) {
-    const randomN = Math.floor(Math.random() * data.length);
-    const randomItem = data[randomN];
-    if (!randomData.includes(randomItem) && randomItem.id !== id) {
-      randomData.push(randomItem);
-    }
-  }
+  const { id } = useParams() as { id: string };
+  const randomData = generateSmallPreview(data, maxNumCardsDesktop, id);
 
   const news = randomData.map((item) => (
     <CardArticlePreview key={item.id} data={item} type="news" route={route} />
